DPC-499: Refactor batch queue name
