+++ dpc-queue/src/main/java/gov/cms/dpc/queue/DistributedBatchQueue.java
@@ -2,7 +2,7 @@ package gov.cms.dpc.queue;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
-import gov.cms.dpc.common.hibernate.queue.DPCQueueManagedSessionFactory;
+import gov.cms.dpc.common.hibernate.DPCManagedSessionFactory;
 import gov.cms.dpc.common.utils.MetricMaker;
 import gov.cms.dpc.queue.annotations.QueueBatchSize;
 import gov.cms.dpc.queue.exceptions.JobQueueFailure;
@@ -26,7 +26,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
 /**
  * Implements a distributed {@link gov.cms.dpc.queue.models.JobQueueBatch} using a Postgres database
@@ -50,7 +49,7 @@ public class DistributedBatchQueue extends JobQueueCommon {
 
     @Inject
     public DistributedBatchQueue(
-            DPCQueueManagedSessionFactory factory,
+            DPCManagedSessionFactory factory,
             @QueueBatchSize int batchSize,
             MetricRegistry metricRegistry
     ) {
@@ -69,12 +68,12 @@ public class DistributedBatchQueue extends JobQueueCommon {
 
     @Override
     protected void submitJobBatches(List<JobQueueBatch> jobBatches) {
-        JobQueueBatch firstBatch = jobBatches.stream().findFirst().orElseThrow(() -> new JobQueueFailure("No job batches to submit"));
+        JobQueueBatch firstBatch = jobBatches.stream().findFirst().orElseThrow();
 
         logger.debug("Adding jobID {} ({} batches) to the queue at {} with for organization {}.",
                 firstBatch.getJobID(),
                 jobBatches.size(),
-                firstBatch.getSubmitTime().orElseThrow(() -> new JobQueueFailure(firstBatch.getJobID(), firstBatch.getBatchID(), "The batches have not been prepared for submission")).format(DateTimeFormatter.ISO_OFFSET_DATE_TIME),
+                firstBatch.getSubmitTime().orElseThrow().format(DateTimeFormatter.ISO_OFFSET_DATE_TIME),
                 firstBatch.getOrgID());
 
         // Persist the batches in postgres
@@ -133,73 +132,53 @@ public class DistributedBatchQueue extends JobQueueCommon {
         }
     }
 
+    @SuppressWarnings("unchecked")
     @Override
-    public Optional<JobQueueBatch> claimBatch(UUID aggregatorID) {
+    public Optional<JobQueueBatch> workBatch(UUID aggregatorID) {
         try (final Session session = this.factory.openSession()) {
             final Transaction tx = session.beginTransaction();
             try {
-                this.restartStuckBatches(session);
-                return this.claimBatchFromDatabase(session, aggregatorID);
+                // Find stuck batches
+                List<String> stuckBatchIDs = session.createSQLQuery("SELECT batch_id FROM job_queue_batch WHERE status = 1 AND update_time > current_timestamp - interval '5 minutes' FOR UPDATE SKIP LOCKED")
+                        .getResultList();
+
+                // Unstick stuck batches
+                if ( stuckBatchIDs != null && !stuckBatchIDs.isEmpty() ) {
+                    final CriteriaBuilder builder = session.getCriteriaBuilder();
+                    final CriteriaQuery<JobQueueBatch> query = builder.createQuery(JobQueueBatch.class);
+                    final Root<JobQueueBatch> root = query.from(JobQueueBatch.class);
+
+                    query.select(root);
+                    query.where(root.get("batchID").in(stuckBatchIDs));
+                    final List<JobQueueBatch> stuckJobList = session.createQuery(query).getResultList();
+
+                    for ( JobQueueBatch stuckJob : stuckJobList ) {
+                        stuckJob.restartBatch();
+                        session.persist(stuckJob);
+                    }
+                }
+
+                // Claim a new batch
+                Optional<String> batchID = session.createSQLQuery("SELECT batch_id FROM job_queue_batch WHERE status = 0 ORDER BY priority ASC, submit_time ASC LIMIT 1 FOR UPDATE SKIP LOCKED")
+                        .uniqueResultOptional()
+                        .map(Object::toString);
+
+                if ( batchID.isPresent() ) {
+                    JobQueueBatch batch = session.get(JobQueueBatch.class, batchID.get());
+                    batch.setRunningStatus(aggregatorID);
+                    session.persist(batch);
+                    session.refresh(batch);
+
+                    final var delay = Duration.between(batch.getStartTime().orElseThrow(), batch.getUpdateTime().orElseThrow());
+                    waitTimer.update(delay.toMillis(), TimeUnit.MILLISECONDS);
+
+                    return Optional.of(batch);
+                } else {
+                    return Optional.empty();
+                }
             } finally {
                 tx.commit();
             }
-        }
-    }
-
-    /**
-     * Looks for any job batches that have stopped processing and are stuck in a running state. Restart those batches
-     * so they can be picked up again.
-     *
-     * @param session - The active database session
-     */
-    @SuppressWarnings("unchecked")
-    private void restartStuckBatches(Session session) {
-        // Find stuck batches
-        List<String> stuckBatchIDs = session.createNativeQuery("SELECT Cast(batch_id as varchar) batch_id FROM job_queue_batch WHERE status = 1 AND update_time < current_timestamp - interval '5 minutes' FOR UPDATE SKIP LOCKED")
-                .getResultList();
-
-        // Unstick stuck batches
-        if ( stuckBatchIDs != null && !stuckBatchIDs.isEmpty() ) {
-            final CriteriaBuilder builder = session.getCriteriaBuilder();
-            final CriteriaQuery<JobQueueBatch> query = builder.createQuery(JobQueueBatch.class);
-            final Root<JobQueueBatch> root = query.from(JobQueueBatch.class);
-
-            query.select(root);
-            query.where(root.get("batchID").in(stuckBatchIDs.stream().map(UUID::fromString).collect(Collectors.toList())));
-            final List<JobQueueBatch> stuckJobList = session.createQuery(query).getResultList();
-
-            for ( JobQueueBatch stuckJob : stuckJobList ) {
-                stuckJob.restartBatch();
-                session.merge(stuckJob);
-            }
-        }
-    }
-
-
-    /**
-     * Claim a new batch to process from the database
-     *
-     * @param session - The active database session
-     * @param aggregatorID - The ID of the aggregator processing the job
-     * @return the claimed job batch
-     */
-    @SuppressWarnings("unchecked")
-    private Optional<JobQueueBatch> claimBatchFromDatabase(Session session, UUID aggregatorID) {
-        // Claim a new batch
-        Optional<String> batchID = session.createNativeQuery("SELECT Cast(batch_id as varchar) batch_id FROM job_queue_batch WHERE status = 0 ORDER BY priority ASC, submit_time ASC LIMIT 1 FOR UPDATE SKIP LOCKED")
-                .uniqueResultOptional();
-
-        if ( batchID.isPresent() ) {
-            JobQueueBatch batch = session.get(JobQueueBatch.class, UUID.fromString(batchID.get()));
-            batch.setRunningStatus(aggregatorID);
-            session.merge(batch);
-
-            final var delay = Duration.between(batch.getStartTime().orElseThrow(), batch.getUpdateTime().orElseThrow());
-            waitTimer.update(delay.toMillis(), TimeUnit.MILLISECONDS);
-
-            return Optional.of(batch);
-        } else {
-            return Optional.empty();
         }
     }
 
@@ -209,7 +188,7 @@ public class DistributedBatchQueue extends JobQueueCommon {
             final Transaction tx = session.beginTransaction();
             try {
                 job.setPausedStatus(aggregatorID);
-                session.merge(job);
+                session.persist(job);
             } finally {
                 tx.commit();
             }
@@ -224,7 +203,8 @@ public class DistributedBatchQueue extends JobQueueCommon {
                 final Optional<OffsetDateTime> lastUpdate = job.getUpdateTime();
 
                 // We just need to persist the job, as any results will be attached to the job and cascade
-                session.merge(job);
+                session.persist(job);
+                session.refresh(job);
 
                 final var delay = Duration.between(lastUpdate.orElseThrow(), job.getUpdateTime().orElseThrow());
                 partialTimer.update(delay.toMillis(), TimeUnit.MILLISECONDS);
@@ -236,15 +216,12 @@ public class DistributedBatchQueue extends JobQueueCommon {
 
     @Override
     public void completeBatch(JobQueueBatch job, UUID aggregatorID) {
-        if ( job == null ) {
-            throw new JobQueueFailure("Empty job passed");
-        }
-
         try (final Session session = this.factory.openSession()) {
             final Transaction tx = session.beginTransaction();
             try {
                 job.setCompletedStatus(aggregatorID);
-                session.merge(job);
+                session.persist(job);
+                session.refresh(job);
 
                 final var delay = Duration.between(job.getStartTime().orElseThrow(), job.getCompleteTime().orElseThrow());
                 successTimer.update(delay.toMillis(), TimeUnit.MILLISECONDS);
@@ -260,7 +237,8 @@ public class DistributedBatchQueue extends JobQueueCommon {
             final Transaction tx = session.beginTransaction();
             try {
                 job.setFailedStatus(aggregatorID);
-                session.merge(job);
+                session.persist(job);
+                session.refresh(job);
 
                 final var delay = Duration.between(job.getStartTime().orElseThrow(), job.getUpdateTime().orElseThrow());
                 failureTimer.update(delay.toMillis(), TimeUnit.MILLISECONDS);
